requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pillow==10.3.0
pydantic==2.7.1
pydantic-settings==2.2.1
cairosvg==2.7.0
